nixos/qemu-vm: Reuse qemu.package for disk image conversion
This may save significant build dependencies, especially when the
test is non-graphical.

To reproduce, set `virtualisation.writableStore = true` in the `gitlab`
test.

    nix-store -qR $(nix-instantiate -A nixosTests.gitlab) | grep gtk

Before this commit, it will show a number of gtk packages.
After this commit, only gtk-doc remains, because the test uses a
much more minimal set of dependencies for its qemu.

It reduced the build closure size from 5399 to 4750 derivations + sources.
Note that the gitlab derivation itself has 3027 paths, so the above decrease
would be quite significant for smaller packages.

Decreasing build closure size reduces the number of rebuilds required
as the closure evolves. In this case, changes in some 650 derivations
+ sources will not require a rerun of the test when they change.